Revert "good point"
This reverts commit 46806b11e6ffa9a1d76e1f8c9b620b877663a7d9.

+++ synthetic_dataset_creation/constellation/constellation.py
@@ -27,7 +27,7 @@ class Constellation:
 
     def generate_pam_constellation(self):
         constellation_points = 2 * (np.arange(self.config.constellation_order) + 1) - 1 - self.config.constellation_order
-        return constellation_points.astype(np.float32)
+        return constellation_points
